Revert "Merge branch 'trunk' into try/persisted-product-form-templates"
This reverts commit 6caf38a93ac8fbf78fb686d2cb80760cc630600c, reversing
changes made to 11dc7e9db3df643a0067bfb23c57323f6d3a2e8f.

+++ plugins/woocommerce/src/Internal/Admin/Logging/FileV2/FileController.php
@@ -4,7 +4,6 @@ declare( strict_types = 1 );
 namespace Automattic\WooCommerce\Internal\Admin\Logging\FileV2;
 
 use Automattic\Jetpack\Constants;
-use Automattic\WooCommerce\Internal\Admin\Logging\Settings;
 use PclZip;
 use WC_Cache_Helper;
 use WP_Error;
@@ -77,6 +76,20 @@ class FileController {
 	private const SEARCH_CACHE_KEY = 'logs_previous_search';
 
 	/**
+	 * The absolute path to the log directory.
+	 *
+	 * @var string
+	 */
+	private $log_directory;
+
+	/**
+	 * Class FileController
+	 */
+	public function __construct() {
+		$this->log_directory = trailingslashit( Constants::get_constant( 'WC_LOG_DIR' ) );
+	}
+
+	/**
 	 * Get the file size limit that determines when to rotate a file.
 	 *
 	 * @return int
@@ -128,7 +141,7 @@ class FileController {
 		}
 
 		if ( ! $file instanceof File ) {
-			$new_path = Settings::get_log_directory() . $this->generate_filename( $source, $time );
+			$new_path = $this->log_directory . $this->generate_filename( $source, $time );
 			$file     = new File( $new_path );
 		}
 
@@ -204,7 +217,7 @@ class FileController {
 		$args = wp_parse_args( $args, self::DEFAULTS_GET_FILES );
 
 		$pattern = $args['source'] . '*.log';
-		$paths   = glob( Settings::get_log_directory() . $pattern );
+		$paths   = glob( $this->log_directory . $pattern );
 
 		if ( false === $paths ) {
 			return new WP_Error(
@@ -319,15 +332,14 @@ class FileController {
 	 * @return File[]
 	 */
 	public function get_files_by_id( array $file_ids ): array {
-		$log_directory = Settings::get_log_directory();
-		$paths         = array();
+		$paths = array();
 
 		foreach ( $file_ids as $file_id ) {
 			// Look for the standard filename format first, which includes a hash.
-			$glob = glob( $log_directory . $file_id . '-*.log' );
+			$glob = glob( $this->log_directory . $file_id . '-*.log' );
 
 			if ( ! $glob ) {
-				$glob = glob( $log_directory . $file_id . '.log' );
+				$glob = glob( $this->log_directory . $file_id . '.log' );
 			}
 
 			if ( is_array( $glob ) ) {
@@ -411,7 +423,7 @@ class FileController {
 
 		$created_pattern = $created ? '-' . gmdate( 'Y-m-d', $created ) . '-' : '';
 
-		$rotation_pattern = Settings::get_log_directory() . $source . $rotations_pattern . $created_pattern . '*.log';
+		$rotation_pattern = $this->log_directory . $source . $rotations_pattern . $created_pattern . '*.log';
 		$rotation_paths   = glob( $rotation_pattern );
 		$rotation_files   = $this->convert_paths_to_objects( $rotation_paths );
 		foreach ( $rotation_files as $rotation_file ) {
@@ -450,7 +462,7 @@ class FileController {
 	 * @return array|WP_Error
 	 */
 	public function get_file_sources() {
-		$paths = glob( Settings::get_log_directory() . '*.log' );
+		$paths = glob( $this->log_directory . '*.log' );
 		if ( false === $paths ) {
 			return new WP_Error(
 				'wc_log_directory_error',
@@ -481,7 +493,10 @@ class FileController {
 
 		$files = $this->get_files_by_id( $file_ids );
 		foreach ( $files as $file ) {
-			$result = $file->delete();
+			$result = false;
+			if ( $file->is_writable() ) {
+				$result = $file->delete();
+			}
 
 			if ( true === $result ) {
 				$deleted ++;
@@ -656,10 +671,10 @@ class FileController {
 	 */
 	public function get_log_directory_size(): int {
 		$bytes = 0;
-		$path  = realpath( Settings::get_log_directory() );
+		$path  = realpath( $this->log_directory );
 
 		if ( wp_is_writable( $path ) ) {
-			$iterator = new \RecursiveIteratorIterator( new \RecursiveDirectoryIterator( $path, \FilesystemIterator::SKIP_DOTS ), \RecursiveIteratorIterator::CATCH_GET_CHILD );
+			$iterator = new \RecursiveIteratorIterator( new \RecursiveDirectoryIterator( $path, \FilesystemIterator::SKIP_DOTS ) );
 
 			foreach ( $iterator as $file ) {
 				$bytes += $file->getSize();
